Code cleanup!

+++ Peon/Source/Game/Entity/Rock.cpp
@@ -17,5 +17,5 @@ void Rock::Update(double deltaTime)
 
 void Rock::Render()
 {
-    m_game->RenderSprite("resource", 2, 0, m_position.x, m_position.y, 32, 32);
+    m_game->RenderSprite("resource", 2, 0, (int)m_position.x, (int)m_position.y, 32, 32);
 }
